fashion-recommender/backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pydantic[email]
requests
motor
dnspython
bcrypt
python-multipart
python-jose[cryptography]
